+++ vitest.config.ts
@@ -8,6 +8,10 @@ export default defineConfig({
   test: {
     environment: 'jsdom',
     globals: true,
+    exclude: [
+      '**/node_modules/**',
+      '**/e2e-tests/**'
+    ],
   },
   resolve: {
     alias: {
